backend/requirements.txt
pydantic-settings>=2.0.0
pyserial==3.5
pysolarmanv5==3.0.6
PyYAML==6.0.3
uModbus==1.0.4
wheel==0.45.1
paho-mqtt==2.1.0
fastapi==0.115.4
uvicorn[standard]==0.32.0
debugpy>=1.8.0

